added cpu util

+++ include/linux_parser.h
@@ -40,7 +40,7 @@ enum CPUStates {
   kGuest_,
   kGuestNice_
 };
-std::vector<unsigned long long> CpuUtilization();
+std::vector<unsigned long> CpuUtilization();
 long Jiffies();
 long ActiveJiffies();
 long ActiveJiffies(int pid);
