+++ include/processor.h
@@ -7,8 +7,8 @@ class Processor {
 
   // TODO: Declare any necessary private members
  private:
-    long double prev_idle{0};
-    unsigned long long prev_total{0};
+    unsigned long prev_idle{0};
+    unsigned long prev_total{0};
 };
 
 #endif
